requirements.txt
Pillow
fpdf
tqdm
pandas